--- src/main/java/org/cactoos/list/AnyOf.java
@@ -1,64 +1,0 @@
-/**
- * The MIT License (MIT)
- *
- * Copyright (c) 2017 Yegor Bugayenko
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.cactoos.list;
-
-import org.cactoos.Scalar;
-
-/**
- * Is {@code true} when any item in the collection is {@code true}.
- *
- * <p>There is no thread-safety guarantee.
- *
- * @author Yegor Bugayenko (yegor256@gmail.com)
- * @version $Id$
- * @since 0.1
- */
-public final class AnyOf implements Scalar<Boolean> {
-
-    /**
-     * Iterable.
-     */
-    private final Iterable<Boolean> iterable;
-
-    /**
-     * Ctor.
-     * @param src Source iterable
-     */
-    public AnyOf(final Iterable<Boolean> src) {
-        this.iterable = src;
-    }
-
-    @Override
-    public Boolean asValue() {
-        boolean success = false;
-        for (final Boolean item : this.iterable) {
-            if (item) {
-                success = true;
-                break;
-            }
-        }
-        return success;
-    }
-
-}
