+++ src/main/java/org/cactoos/list/FilteredIterator.java
@@ -100,7 +100,7 @@ public final class FilteredIterator<X> implements Iterator<X> {
     public X next() {
         if (!this.hasNext()) {
             throw new NoSuchElementException(
-                "No more elements with fits the condition."
+                "No more elements that fit the condition"
             );
         }
         return this.buffer.poll();
@@ -109,7 +109,7 @@ public final class FilteredIterator<X> implements Iterator<X> {
     @Override
     public void remove() {
         throw new UnsupportedOperationException(
-            "FilteredIterator does not support remove Operation"
+            "#remove() is not supported"
         );
     }
